+++ src/datastore/async_methods/destroyAll.js
@@ -15,7 +15,7 @@ function destroyAll(resourceName, params, options) {
       reject(new DSErrors.IA('"options" must be an object!'));
     } else {
       if (!('notify' in options)) {
-        options.notify = true;
+        options.notify = definition.notify;
       }
       if (!('eagerEject' in options)) {
         options.eagerEject = definition.eagerEject;
@@ -23,23 +23,23 @@ function destroyAll(resourceName, params, options) {
       resolve();
     }
   }).then(function () {
-      var func = options.beforeDestroy ? promisify(options.beforeDestroy) : definition.beforeDestroy;
+      var func = options.beforeDestroy ? DSUtils.promisify(options.beforeDestroy) : definition.beforeDestroy;
       toEject = _this.defaults.defaultFilter.call(_this, resourceName, params);
       return func(resourceName, toEject);
     }).then(function () {
       if (options.notify) {
-        _this.notify(definition, 'beforeDestroy', toEject);
+        _this.emit(definition, 'beforeDestroy', toEject);
       }
       if (options.eagerEject) {
         ejected = _this.ejectAll(resourceName, params);
       }
       return _this.getAdapter(definition, options).destroyAll(definition, params, options);
     }).then(function () {
-      var func = options.afterDestroy ? promisify(options.afterDestroy) : definition.afterDestroy;
+      var func = options.afterDestroy ? DSUtils.promisify(options.afterDestroy) : definition.afterDestroy;
       return func(resourceName, toEject);
     }).then(function () {
       if (options.notify) {
-        _this.notify(definition, 'afterDestroy', toEject);
+        _this.emit(definition, 'afterDestroy', toEject);
       }
       return ejected || _this.ejectAll(resourceName, params);
     }).catch(function (err) {